Upgrade to latest webpack (5.75.0)
When using Node 17+, the webpack commands fail with the error
`ERR_OSSL_EVP_UNSUPPORTED`. The solution given by [miken32 on StackOverflow][1]
does not work with VS Code since [Electron allows only a subset of NODE_OPTIONS][2].
We should instead upgrade to the latest webpack which works properly with
OpenSSL 3.

The webpack upgrade requires a few other package changes so that everything
still builds and runs.

[1]: https://stackoverflow.com/questions/70582072/npm-run-fails-with-err-ossl-evp-unsupported/70582385#70582385
[2]: https://github.com/electron/electron/blob/main/docs/api/environment-variables.md#node_options