Make it possible to run multiple async VS Code commands
[We're currently using a `forEach` loop][1]  to run all VS Code commands
that we pass in a single Code Tour step.

After testing this with our [codespaces-codeql repo][2], we've found
that if multiple commands are async, only the first one gets triggered
successfully.

This is because `forEach` loops [won't await promises][3] and will
instead throw them away.

If we want to run our commands in sequence, it's recommended that we use
a modern `for ... of` loop instead.

[1]: https://github.com/microsoft/codetour/blob/e7018e228c64f9dde061af56217927159cd42950/src/player/index.ts#L400
[2]: https://github.com/github/codespaces-codeql
[3]: https://stackoverflow.com/questions/37576685/using-async-await-with-a-foreach-loop

Co-authored-by: Aditya Sharad <adityasharad@github.com>